requirements.txt
Contexto~=0.1.0
numpy~=1.20.3
matplotlib~=3.3.1
scipy~=1.6.3
sklearn~=0.24.2
scikit-learn~=0.24.2
pandas~=1.2
requests~=2.25.1
flask~=2.0.1
logging~=0.4.9.5
mysql-connector-python~=8.0.25